% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretocounts.R
\name{dparetocounts}
\alias{dparetocounts}
\title{Bounded power law probability density function}
\usage{
dparetocounts(x, mu, vreal2, vreal3)
}
\arguments{
\item{x}{body size value}

\item{mu}{vector of lambda (the power law exponent)}

\item{vreal2}{xmin: the minimum body size of the sample or the minimum possible body size}

\item{vreal3}{xmax: the maximum body size of the sample or the maximum possible body size}
}
\value{
a numeric vector of the value of the pdf given values of x, mu, xmin, and xmax.
}
\description{
Bounded power law probability density function
}
\examples{
dparetocounts(x = 2, mu = -1.5, vreal2 = 1, vreal3 = 1000)
}
