% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretocounts.R
\name{log_lik_paretocounts}
\alias{log_lik_paretocounts}
\title{Generate custom log-likelihood for Stan}
\usage{
log_lik_paretocounts(i, prep)
}
\arguments{
\item{i}{see documentation from `get_dpar` within `brms`}

\item{prep}{see documentation from `get_dpar` within `brms`}
}
\value{
NA
}
\description{
This function is not called directly. It is only used to set up the log likelihood for
later use in `brms` via the family option within `brm()`
}
\examples{
NA
}
