% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretocounts.R, R/paretocounts_original.R
\name{paretocounts}
\alias{paretocounts}
\title{Create custom paretocounts family}
\usage{
paretocounts()

paretocounts()
}
\value{
NA

NA
}
\description{
This function is not called directly. It is only used to create the paretocounts() family option
in `brms`

This function is not called directly. It is only used to create the paretocounts() family option
in `brms`
}
\examples{
NA
NA
}
