% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretocounts.R
\name{paretocounts_lpdf_temp}
\alias{paretocounts_lpdf_temp}
\title{Create paretocounts log posterior density function}
\usage{
paretocounts_lpdf_temp(Y, mu, vreal1, vreal2, vreal3)
}
\arguments{
\item{Y}{vector of data (individual body sizes)}

\item{mu}{exponent of the bounded power law}

\item{vreal1}{counts for each body size (numeric of integer)}

\item{vreal2}{xmin: the minimum body size of the sample or the minimum possible body size}

\item{vreal3}{xmax: the maximum body size of the sample or the maximum possible body size}
}
\value{
temporary function that is later vectorized using vectorize(paretocounts_lpdf_temp)
}
\description{
This function is not called directly. It is only used to create the paretocounts() family option
in `brms`
}
\examples{
NA
}
