% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretocounts.R
\name{posterior_predict_paretocounts}
\alias{posterior_predict_paretocounts}
\title{Arrange data for posterior prediction}
\usage{
posterior_predict_paretocounts(i, prep, ...)
}
\arguments{
\item{i}{see documentation from `get_dpar` within `brms`}

\item{prep}{see documentation from `get_dpar` within `brms`}

\item{...}{}
}
\value{
NA
}
\description{
This function is not called directly.
}
\examples{
NA
}
