% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretocounts.R
\name{pparetocounts}
\alias{pparetocounts}
\title{Bounded power law cumulative probability density function}
\usage{
pparetocounts(x, xmin, xmax, lambda)
}
\arguments{
\item{x}{body size value}

\item{xmin}{xmin: the minimum body size of the sample or the minimum possible body size}

\item{xmax}{xmax: the maximum body size of the sample or the maximum possible body size}

\item{lambda}{vector of lambda (the power law exponent)}
}
\value{
a numeric vector of the probability of x given lambda, xmin, and xmax.
}
\description{
Bounded power law cumulative probability density function
}
\examples{
pparetocounts(x = 2, xmin = 1, xmax = 1000, lambda = -2)
}
