% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretocounts.R
\name{rparetocounts}
\alias{rparetocounts}
\title{Simulate data from a bounded power law}
\usage{
rparetocounts(n = 300, mu = -1.2, vreal2 = 1, vreal3 = 1000)
}
\arguments{
\item{n}{number of observations}

\item{mu}{vector of lambda (the power law exponent)}

\item{vreal2}{xmin: the minimum body size of the sample or the minimum possible body size}

\item{vreal3}{xmax: the maximum body size of the sample or the maximum possible body size}
}
\value{
a numeric vector
}
\description{
Simulate data from a bounded power law
}
\examples{
rparetocounts(n = 100, mu = -1.5, vreal2 = 1, vreal3 = 2000)
}
