% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretocounts.R, R/paretocounts_original.R
\name{dparetocounts}
\alias{dparetocounts}
\title{Bounded power law probability density function}
\usage{
dparetocounts(x, mu, vreal2, vreal3)

dparetocounts(x, mu, vreal2, vreal3)
}
\arguments{
\item{x}{body size value}

\item{mu}{vector of lambda (the power law exponent)}

\item{vreal2}{xmin: the minimum body size of the sample or the minimum possible body size}

\item{vreal3}{xmax: the maximum body size of the sample or the maximum possible body size}
}
\value{
a numeric vector of the value of the pdf given values of x, mu, xmin, and xmax.

a numeric vector of the value of the pdf given values of x, mu, xmin, and xmax.
}
\description{
Bounded power law probability density function

Bounded power law probability density function
}
\examples{
xmin = 1
xmax = 1000
lambda = -1.5

dparetocounts(x = 2, mu = lambda, vreal2 = xmin, vreal3 = xmax)
xmin = 1
xmax = 1000
lambda = -1.5

dparetocounts(x = 2, mu = lambda, vreal2 = xmin, vreal3 = xmax)
}
