% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretocounts.R
\name{dparetocounts}
\alias{dparetocounts}
\title{Bounded power law probability density function}
\usage{
dparetocounts(x, lambda, xmin, xmax)
}
\arguments{
\item{x}{body size value}

\item{lambda}{vector of lambda (the power law exponent)}

\item{xmin}{xmin: the minimum body size of the sample or the minimum possible body size}

\item{xmax}{xmax: the maximum body size of the sample or the maximum possible body size}
}
\value{
a numeric vector of the value of the pdf given values of x, lambda, xmin, and xmax.
}
\description{
Bounded power law probability density function
}
\examples{
xmin = 1
xmax = 1000
lambda = -1.5

dparetocounts(x = 2, lambda = lambda, xmin = xmin, xmax = xmax)
}
