% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretocounts.R
\name{rparetocounts}
\alias{rparetocounts}
\title{Simulate data from a bounded power law
Simulates data from a bounded power law. This is basically a vectorized version of the rPLB() function
in Edwards et al. (2017) `sizeSpectra` package. The argument names are required to match the default
arguments for custom likelihoods in `brms`.}
\usage{
rparetocounts(n = 300, lambda = -1.2, xmin = 1, xmax = 1000)
}
\arguments{
\item{n}{number of observations}

\item{lambda}{vector of lambda (the power law exponent)}

\item{xmin}{xmin: the minimum body size of the sample or the minimum possible body size}

\item{xmax}{xmax: the maximum body size of the sample or the maximum possible body size}
}
\value{
a numeric vector
}
\description{
Simulate data from a bounded power law
Simulates data from a bounded power law. This is basically a vectorized version of the rPLB() function
in Edwards et al. (2017) `sizeSpectra` package. The argument names are required to match the default
arguments for custom likelihoods in `brms`.
}
\examples{
rparetocounts(n = 100, lambda = -1.5, xmin = 1, xmax = 2000)
}
